+++ GeometryLibrary/src/Cone3D.cpp
@@ -7,7 +7,7 @@ namespace Fierce {
 		addVertex3D(vertices, 0.0f, 0.0f, 0.5f);
 	}
 
-	void Cone3D::getIndices(std::vector<unsigned int>& indices, int numPoints, float angle, float innerRadius, int numRings) {
+	void Cone3D::getIndices(std::vector<uint16_t>& indices, int numPoints, float angle, float innerRadius, int numRings) {
 		addCircleIndices(indices, numPoints, angle, 0, 1, false);
 		addCircleIndices(indices, numPoints, angle, numPoints + 1, 1, true);
